requirements.txt
boto3==1.9.105
botocore==1.12.105
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
Werkzeug==0.14.1
isodate>=0.5.0
xmlsec>=0.3.1
defusedxml==0.5.0
flask-dynamodb-sessions==0.1.4
flask-cors==3.0.7
pyjwt==1.7.1
cryptography==2.6.1